pyproject.toml
[tool.poetry]
name = "thesisv2"
version = "0.1.0"
description = ""
authors = ["wkhoury <wkhoury>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.5.3"
numpy = "^1.24.2"
faker-music = "^0.4"
configparser = "^5.3.0"
pyyaml = "^6.0"
jupyter = "^1.0.0"
tqdm = "^4.66.1"
scikit-learn = "^1.3.0"
matplotlib = "^3.8.0"
seaborn = "^0.13.2"
pytest = "^8.1.1"
xgboost = "^2.1.0"
faker = "^33.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
